requirements.txt
langchain_google_genai
streamlit
langchain
langchain-ollama
python-dotenv
streamlit